Support for Docker
Just a test at this point. 

Docker Watcher disclaimer:

This is a rough first start to getting a docker container for Watcher. I have managed to get it running but have not managed to list volumes to change directory listing for the config file and other directories. This is a first every attempt. Any input you have in getting volumes listed so they can be set in docker would be appreciated. It seems if you restart the docker container or restart within Watcher it saves the config file. I cannot change the path to where the config file is at this stage.