Quick-add position changes based on container
Because the height of the movie `li` container is variable based on title length and poster, calculating the absolute position from the bottom for an overlay element can be problematic.  Basing the position based on the top, which is static, means that the quick-add overlay will always be in the same place, regardless of the height of the container.  The position `top: 15em;` works well for this specific application.

Orig: http://imgur.com/a/RHkpS
Edit: http://imgur.com/a/JmEjS