Add CSS Theming
Towards the bottom of the @settings_page section, a drop-down has been added that parses the comma delimited theming options from the config file.  Since it appears that the config file is re-written upon saving, the theme list has been added to a hidden field so that those options are preserved throughout the editing/saving process.