Reduce rate of tmdb api calls
We get 40 requests that regenerate at 4/second. When regenerating our tokens at 3.5/s I'd still run out and get 429 responses. This reduces the token capacity to 35 so we have a bit more buffer ony sends a request if we have more than two full tokens.

I was still able to run an import of 370 movies in less than 5 minutes, and that is speedy enough.