Wait for list iteration to finish before firing post request in import
Oh my hatred for javascript is a boiling cauldron of rage and vitriol.

This forces the script to wait until every $.each() loop is finished before sending the data back to the server. Why javascript doesn't wait for a loop to finish without explicit instructions to do so is something I will never understand. Stupid language. Stupid stupid language.

But I was able to scan and import ~370 movies with no trouble, so as long as I don't have to touch this script again I'll be happy.