Use streaming output to process imports
Rather than sending an ajax request and waiting for the result I'm now yielding each file's result back to the browser in real-time. This does a few things.

* Lets the user know that progress is being made and at what rate.
* Prevents the browser from abandoning the ajax request for a long-running scan.

The interface is a bit more clumsy than I'd like it to be, but I'm not going to worry about that until it actually works the way I want.