Better handling of title comparisons
Use string.punctuation as a filter to remove any punctuation characters before url quoting. This way words won't get split into partial words that are harder to find a match on.

I'm not 100% sure the urllib2.quote() call is neccesary any more, but it costs very little and can't hurt.